requirements_backup.txt
streamlit
google-genai
borax
fpdf2
iztro-py
gspread
google-auth
